full upload - finalised
A quiz game built in Python intended to allow some casual conversation and insight into personality and interests during presentations/employment pitches. 
Multiple difficulty options are available, functions and dictionaries for which are contained within their own independent modules. 
A custom exceptions module has been created and imported into the main file as well, however this isn't actually used. It was created more as a test of knowledge rather than because it was needed. 
Plan of action is included at the top of the main file, as well as various comments throughout the main file and linked modules. The "Better Comments" extension for Visual Studio Code wil be needed in order to view formatting for these correctly.